Add input/output queue and filter out hardlink candidates if contents are emitted
We can do a bit better with preventing excessive events by batching all
incoming change events in a macro tick. In each tick, we'll process the
events, and filter them as usual, but also mark hardlink candidates, as
before.

However, we now omit hardlink candidates, if any sub-files/directories
are in the batch, which indicates usually that `fs.watch` will deliver
children and the candidate is not a clone/hardlink/copy-on-write directory.

We then also batch the outputs and process all pending ones in the same
tick. Due to the async/await logic, we usually won't process inputs and
outputs in the same tick.