Add script to generate .d.ts files (#1562)
Summary:
Add a script to generate `.d.ts` files from Flow source using `flow-api-translator`, for source files matching provided globs.

Currently we have some `.d.ts` files checked in, but several are out of date and there are gaps.

The plan would be that we incrementally expand these globs to cover the whole repository (or at least a stable set of whole packages) and *then delete the checked-in `types`* in favour of generating them as a prepublish step for NPM only.

Pull Request resolved: https://github.com/facebook/metro/pull/1562

Changelog: Internal

Reviewed By: huntie

Differential Revision: D80535987

fbshipit-source-id: e2449663d17a3d80678edbfd4b21081bc24e6e54