remove support for relative urls when parsing bundle options from url
Summary:
Changelog: [Internal] Remove support for relative urls when parsing bundle options from url. This seems to have not been used.

Working on `parseBundleOptionsFromBundleRequestUrl`, I noticed we handle a use-case that is not clear why would we support.

It complicates the logic of that very complicated function without a sufficiently good reason, besides being a left-over from an older HMR logic.

While we introduce some semi-breaking-changes to how urls are handled, it's a good opportunity to make the function stricter in terms of what it expects to receive as an input.

Reviewed By: robhogan

Differential Revision: D79809398

fbshipit-source-id: 3b73e189ef163d1a965a71883c83e826b8418f0f