Fix population of default options in `runBuild` API
Summary:
The `runBuild` API and `shared/output/bundle.js` abstraction currently do not correctly pass Metro defaults to `Server.build()` and `Server.getAssets()` where overrides are omitted.

In particular, `customResolverOptions`, `customTransformOptions`, and `unstable_transformProfile`, if omitted, make it all the way to the transformer as `undefined`, contrary to Flow types.

Besides being unsafe (though as it happens, we don't crash), this creates a cache key divergence between `runBuild` builds and Metro server builds.

Changelog: Internal - necessary but not yet sufficient to align cache keys

## Flow
This happens because Flow allows a prop be be present with value `undefined` wherever it allows the prop to be omitted, but assumes the prop is either omitted or populated with its specified type in contexts such as object spread.

This doesn't reflect runtime behaviour, and creates a safety issue in the common pattern that overrides are spread across defaults.

## This diff
Adds a utlity type `SafeOptionalProps` to turn `{foo?: type}` into `{foo?: void | type}`, which may be used to force us to handle `undefined` props.

Reviewed By: vzaidman

Differential Revision: D81577054

fbshipit-source-id: 89de6bf3280ba698cc443554ab3823eb0080fba6