respect "x-forwarded-host" and "x-forwarded-proto" when calculating asset url for asset requests from Metro to allow making request through proxy that re-writes the host and protocol
Summary:
X-link: https://github.com/facebook/react-native/pull/53716

Changelog: [Feature] Respect "x-forwarded-host" and "x-forwarded-proto" when calculating asset url for asset requests from Metro to allow making request through proxy that re-writes the host and protocol.

When calculating the url for an asset (such as `bundle.js.bundle` and `bundle.js.map`) requested from Metro through a proxy, to be added as `//# sourceURL` and `//# sourceMappingURL` at the bottom of files, the Host header is currently used to calculate where that asset lives. These two comments are used to resolve the source and source map files for a bundle.

However the header "host" may change if the request is made through a proxy. In these cases, a standard practice is to use the ["x-forwarded-host"](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/X-Forwarded-Host) and ["x-forwarded-proto"](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/X-Forwarded-Proto) header to signify what host the request originally was used before the proxy rewrite to obtain the asset, which is the one we want to use to calculate these comments.

Reviewed By: robhogan

Differential Revision: D81774209

fbshipit-source-id: d81f8684fcaf2b55692efb0e50e562bd8cf0cc90