loadConfig: Don't obscure (syntax, runtime, etc.) errors loading config
Summary:
In https://github.com/facebook/metro/pull/1560, I added some logic to catch errors and throw more helpful ones (given the config path in a new message, and a specific note for the case that a user was trying to use TypeScript config on an older Node.js version).

I preserved the original error in `error.cause`. However, the integration with `react-native-community/cli` discards the cause. This meant other run-of-the-mill errors in `metro.config.js` that prevented the module being evaluated, like a typo in a variable name, or a syntax error, were obscured.

Instead, preserve the original error and stack, and just add a bit to the message before re-throwing it.

Changelog:
```
 - **[Fix]** Don't obscure errors parsing/evaluating metro.config.js
```

Reviewed By: motiz88

Differential Revision: D83559678

fbshipit-source-id: 82c9425463ad1c7833a14efce513d6ed2a657283