metro-file-map: Only expose a plugin's own data to plugins
Summary:
Refactor `metro-file-map`'s plugin API so that plugin methods are provided with the plugin's data (as extracted by the worker, and soon to be the plugin's own worker) and not the entire `FileMetadata` tuple.

This is a little safer and cleaner, but importantly it's a step towards plugins being dynamically configurable, rather than at a fixed position in a tuple.

Changelog: Internal

Reviewed By: huntie

Differential Revision: D83364233

fbshipit-source-id: 33c1ff615729495c56660fd698eb6e3d38bf2635