File map: Use nulls instead of empty strings in non-Haste metadata storage
Summary:
Remove a footgun in handling of stored Haste IDs in `metro-file-map`, where we use empty strings to represent "not Haste".

Use `null` instead, which is also more efficient for memory and serialisation. (Previously, we preferred empty strings for JSON serialisation - 2 characters instead of 4 - but with v8 a null is one byte instead of two)

Changelog: Internal

Reviewed By: huntie

Differential Revision: D83361951

fbshipit-source-id: 8579efaadfd521c2d8f76a8ee1748c62592b64e1