Remove H.ID in favour of generic H.PLUGINDATA start index
Summary:
Make `metro-file-map`'s `FileMetadata` an inexact tuple, and replace the uses of `H.ID`, which previously concretely referred to Haste name in a fixed location, with `H.PLUGINDATA`, which refers to the index of the first configured plugin (temporarily special-cased to Haste because plugins are not yet configurable).

This decouples Haste from the internal data structure, and allows for additional plugins to sit alongside.

Changelog: Internal

Reviewed By: huntie

Differential Revision: D76429293

fbshipit-source-id: 4799faa2d0cab51ac16eb080b76d2d88bc788911