metro-file-map: Fix `hasteImplModulePath` as a worker setup argument, use `computeHaste` to enable/disable Haste per-file
Summary:
Currently, `hasteImplModulePath`, used to enable Haste modules, is an optional argument to the `metro-file-map` constructor - ie, it takes one fixed value for the the lifetime of an instance.

However, we currently pass it, or `null`, each time to the worker when processing files, because
 - Previously we had no notion of a "setup argument" - workers needed to be passed all config with every job
 - Currently we use string vs null to determine whether a worker should process a file for Haste at all - eg, we pass `hasteImplModulePath: null` to workers when processing a `node_modules` file.

This diff
 1. Moves `hasteImplModulePath` to a setup arg, which simplifies enforcing that it's invariant.
 2. Uses `computeHaste: boolean` to replace "pass null haste impl (or `enableHastePackages: false`) to disable Haste computation for this file"

This is an intermediate step on the way to introducing generic filters for plugins.

Reviewed By: huntie

Differential Revision: D88157825

fbshipit-source-id: 19794abe7a4e72b2b08a614911693ccd9ca92f53