Filter no-op jobs for metro-file-map workers before calculating number of workers
Summary:
Currently, `metro-file-map` spins up a number of child workers to help process files, based on the number of added/modified files (excluding `node_modules`) since the last cache.

This isn't optimal - in the common case that Haste is not in use and SHA1 is calculated lazily, we won't actually invoke the worker (`#getWorkerInput` returns `null`), so we've spun up more workers than we need.

Instead, run `#getWorkerInput` for each file to maybe process, filter out those with nothing to do, and calculate the number of workers based on the number of actual tasks we have for workers.

Changelog:
```
 - **[Performance]**: Don't unnecessarily spin up many file map workers when configured with lazy SHA1 and no Haste (default)
```

Reviewed By: huntie

Differential Revision: D84835964

fbshipit-source-id: 5eb52fd2094f480a2fbdb895517e17961c87a9a8