File map: Don't skip hashing package.json if eagerly hashing other files
Summary:
Remove this early filtering of `package.json` files from `metro-file-map` processing.

This bailout exists because up to now, the only reason for `metro-file-map` to visit a package.json file was to process a Haste package. However, with generic plugins, that's not necessarily the case, and this filter is too early. (In particular, we'll need to change this to allow the resolver to read `package.json` files through `metro-file-map`, as it should)

The effect of removing it is that we'll visit these files to compute their SHA1 (if lazy SHA1 is disabled) even if Haste packages are not enabled.

This is *slightly* wasteful on a very small subset of files, but not typically an issue because SHA1 is lazy in OSS (so no change) and from Saved State internally (so the cost is borne in CI).

Reviewed By: huntie

Differential Revision: D86515170

fbshipit-source-id: 5666b108c381b6d5dec425f6c8f39b0d526ef630