metro-file-map: Delegate Haste filtering to Haste plugin, decouple FileProcessor from Haste
Summary:
Introduce `filter` functions to `FileMapPluginWorker`, allowing us to encapsulate the current `computeHaste` logic in `FileProcessor` entirely within `HastePlugin`.

In place of `computeHaste`, we now have the more generic `pluginsToRun`, a list of plugin indices whose filter function passes for the current file.

The main worker will now only call a plugin's `processFile` for each of the `pluginsToRun`. The `pluginData` it returns corresponds to only plugins it was asked to run.

The *almost* completes encapsulating of all Haste-related logic from `metro-file-map` into the `HastePlugin`, except in arguments to the `FileMap` constructor - we'll fix that next. More importantly, it's now clear that this abstraction is capable of handling any Haste-analogous custom plugin - eg deriving data from a file's contents, persisting it and repopulating it, with configurable behaviour and filtering.

Changelog: Internal

Reviewed By: huntie

Differential Revision: D88070695

fbshipit-source-id: 19163a0035dc9b32ec218dfc6e823cdd50ba4bbe