Improve type safety of `parseJsonBody` and `/symbolicate` endpoint, respond 400 vs 500 on bad requests
Summary:
Safely type the output of internal `parseJsonBody`, and add the necessary validation to `/symbolicate` to assert the previously assumed type.

This is not intended to *strengthen* validation - ie it should be non-breaking. The intention is just to make assumptions more explicit. As a semver fix, we use the more appropriate 400 status code for request issues.

The motivation for this is mostly to remove the `$FlowFixMe` from `parseJsonBody`, which blocks TS type generation, but I thought I'd beef up this validation as a drive-by.

Changelog:
```
 - **[Fix]**: `/symbolicate` endpoint responds with 400 (not 500) on validation errors
```

Reviewed By: vzaidman

Differential Revision: D89844824

fbshipit-source-id: ac1bde2916ddcb30aadb2b719ac9c0f1eb03671c